requirements.txt
python-telegram-bot==13.15
fpdf==1.7.2
python-dotenv==1.0.1
telegram
flask